+++ NEWS.md
@@ -3,8 +3,8 @@
 - New feature, argument indexing: `trace[1, 2, :vec]` gets the `:vec` argument of the second call of the first call.
 - Similar to `vec[end]`, `stacktrace[bottom-1]` returns the second-to-last call in a
 stack trace (see `@stacktrace`)
-- `sum(fn, trace)` and `all(fn, trace)` were added. The passed functions should accept
-a `Trace` object.
+- `sum(fn, trace)` and `all(fn, trace)` apply `fn(::Trace)` to each argument.
+- `code_warntype(::Trace), code_llvm(::Trace), ...` return the corresponding code.
 - New function: `highlight(pred, trace)`. See `?highlight`.
 - Profiling with `measure` was significantly improved, with a new option `explore_worst`
 added for quickly surverying large traces. See `?measure` and the user manual.